Create an Admin Users endpoint in the broker
