+++ datavault-broker/src/main/webapp/WEB-INF/datavault-broker-servlet.xml
@@ -84,7 +84,7 @@
         <property name="eventService" ref="eventService" />
     </bean>
 
-    <bean id="adminController" class="org.datavaultplatform.broker.controllers.AdminController">
+    <bean id="adminController" class="org.datavaultplatform.broker.controllers.admin.AdminController">
         <property name="vaultsService" ref="vaultsService" />
         <property name="depositsService" ref="depositsService" />
         <property name="retrievesService" ref="retrievesService" />
@@ -122,6 +122,10 @@
     <bean id="notifyController" class="org.datavaultplatform.broker.controllers.NotifyController">
         <property name="eventService" ref="eventService" />
         <property name="clientsService" ref="clientsService" />
+        <property name="usersService" ref="usersService" />
+    </bean>
+
+    <bean id="adminUsersController" class="org.datavaultplatform.broker.controllers.admin.AdminUsersController">
         <property name="usersService" ref="usersService" />
     </bean>
